removed specialized code out from the core
