+++ models/User.php
@@ -9,8 +9,8 @@
 
 namespace app\models;
 
-use app\core\Application;
 use app\core\UserModel;
+use DateTime;
 
 class User extends UserModel
 {
@@ -133,6 +133,57 @@ class User extends UserModel
     public function getUserByNIC($nic)
     {
         return (new User)->findOne(User::class, ['nic' => $nic]);
+    }
+    public function extractFromNic($nic): array
+    {
+        $dates = [31, 29, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
+
+        //nic length
+        $nicLength = strlen($nic);
+
+        $daysOld = (int)substr($nic, 2, 3);
+        $daysNew = (int)substr($nic, 4, 3);
+
+        //checking for v at last for old nic
+        $checkV = substr_compare($nic, "v", -1) || substr_compare($nic, "V", -1);
+
+        //check for first digit in new nic
+        $checkOne = substr($nic, 0, 1) === '1';
+        $checkTwo = substr($nic, 0, 1) === '2';
+
+        $isOld = true;
+        //validating
+        if ($nicLength == 10 && $checkV && ($daysOld <= 366 || ($daysOld >= 501 && $daysOld <= 866))) {
+            //this is an old nic
+            $isOld = true;
+        } else if ($nicLength == 12 && ($checkOne || $checkTwo) && ($daysNew <= 366 || ($daysNew >= 501 && $daysNew <= 866))) {
+            //this is an old nic
+            $isOld = false;
+        }
+
+        $year = ($isOld) ? 1900 + (int)substr($nic, 0, 2) : (int)substr($nic, 0, 4);
+        $threeDigitsForDays = ($isOld) ? $daysOld : $daysNew;
+
+        //Validate gender
+        $gender = "Male";
+        if ($threeDigitsForDays > 500) {
+            $threeDigitsForDays -= 500;
+            $gender = "Female"; //If day value > 500 it means NIC owner is a female.
+        }
+
+        $total = 0;
+        $date = $month = 0;
+        for ($i = 0; $i <= sizeof($dates); $i++) {
+            $total += $dates[$i];
+            if ($threeDigitsForDays <= $total) {
+                $month = $i + 1; //Get the month
+                $date = $threeDigitsForDays - ($total - $dates[$i]); //Get the day
+                break;
+            }
+        }
+        $date = new DateTime("$year-$month-$date");
+        $dob = $date->format('Y-m-d');
+        return ['dob' => $dob, 'gender' => $gender];
     }
 
 }
